+++ packages/anthro/anthro/utils/igo_land_use.py
@@ -6,7 +6,6 @@
 """
 import sqlite3
 from rscommons import Logger
-from rscommons.database import write_db_attributes
 
 
 def calculate_land_use(database: str, windows: dict):
